BUG-E based Guidance System, Arrow that points you towards teleporter
After traveling a certain amount of distance (1000 currently) BUG-E informs you he is detecting a large amount of Essence in this direction, with a directional arrow pointing toward the teleporter, however far away it may be. Ideally this will show up every 3-5 minutes of playing to keep you on track if you get lost.