BUG-E external dialogue trigger system overhaul
the old system of a collider triggering dialogue has been replaced with the OptionalDialogueZone script. Placing this script on any object in the scene makes it able to trigger optional dialogue, with inspector given values. BUG-E displays an exclaimation mark when in an optional dialogue zone. Pressing F starts the dialogue sequence.